pyproject.toml
[tool.poetry]
name = "tldlbot"
version = "0.1.0"
description = ""
authors = ["Max Ries <max.ries113@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.6"
requests = "^2.25.1"
ibm-watson = "^5.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tldlbot = "tldlbot.main:main"